feat: 2D ControlView 물체 이동 기능 완료

+++ C6trier/Source/Model/ShapeModel.swift
@@ -4,7 +4,7 @@ struct Shape2DModel: ShapeModelProtocol {
     let name: String
     var color: Color
     var radius: CGFloat
-    var position: SIMD2<CGFloat>
+    var position: SIMD2<Float>
     let shapeType: ShapeType
     
     var id: String { name }
@@ -14,7 +14,7 @@ struct Shape3DModel: ShapeModelProtocol {
     let name: String
     var color: Color
     var radius: CGFloat
-    var position: SIMD3<CGFloat>
+    var position: SIMD3<Float>
     let shapeType: ShapeType
     
     var id: String { name }
